+++ agent/CHANGELOG.rst
@@ -6,6 +6,7 @@ This file keeps track of all notable changes to license-manager-agent
 
 Unreleased
 ----------
+* Improve function to get bookings sum from backend
 
 3.0.3 -- 2023-08-17
 -------------------